点赞相关uint改成int64

+++ internal/service/favorite_service/check_favorite.go
@@ -3,7 +3,7 @@ package favorite_service
 import "douyin-lite/internal/entity"
 
 // Check_Favorite 查询某用户是否点赞某视频
-func Check_Favorite(uid uint, vid uint) bool {
+func Check_Favorite(uid int64, vid int64) bool {
 	isFavorite, err := entity.NewFavoriteDaoInstance().Query_Check_Favorite(uid, vid)
 	if err != nil {
 		return false
